1.15

+++ src/main/java/net/querz/mcaselector/io/job/ChunkImporter.java
@@ -1,7 +1,13 @@
-package net.querz.mcaselector.io;
+package net.querz.mcaselector.io.job;
 
 import net.querz.mcaselector.Config;
 import net.querz.mcaselector.debug.Debug;
+import net.querz.mcaselector.io.ByteArrayPointer;
+import net.querz.mcaselector.io.FileHelper;
+import net.querz.mcaselector.io.MCAFilePipe;
+import net.querz.mcaselector.io.RegionDirectories;
+import net.querz.mcaselector.io.SelectionData;
+import net.querz.mcaselector.io.WorldDirectories;
 import net.querz.mcaselector.io.mca.EntitiesMCAFile;
 import net.querz.mcaselector.io.mca.PoiMCAFile;
 import net.querz.mcaselector.io.mca.Region;
@@ -185,17 +191,19 @@ public class ChunkImporter {
 		@Override
 		public void execute() {
 
-			boolean allCopied = true;
-
 			// try to copy files directly if there is no offset, no selection and the target file does not exist
 			if (offset.getX() == 0 && offset.getZ() == 0 && (selection == null || selection.size() == 0)) {
+				boolean allCopied = true;
+
 				if (!getRegionDirectories().getRegion().exists()) {
 					//if the entire mca file doesn't exist, just copy it over
 					File source = new File(sourceDirs.getRegion(), getRegionDirectories().getLocationAsFileName());
-					try {
-						Files.copy(source.toPath(), getRegionDirectories().getRegion().toPath());
-					} catch (IOException ex) {
-						Debug.dumpException(String.format("failed to copy file %s to %s", source, getRegionDirectories().getRegion()), ex);
+					if (source.exists()) {
+						try {
+							Files.copy(source.toPath(), getRegionDirectories().getRegion().toPath());
+						} catch (IOException ex) {
+							Debug.dumpException(String.format("failed to copy file %s to %s", source, getRegionDirectories().getRegion()), ex);
+						}
 					}
 				} else {
 					allCopied = false;
@@ -203,10 +211,12 @@ public class ChunkImporter {
 
 				if (!getRegionDirectories().getPoi().exists() && sourceDirs.getPoi() != null) {
 					File source = new File(sourceDirs.getPoi(), getRegionDirectories().getLocationAsFileName());
-					try {
-						Files.copy(source.toPath(), getRegionDirectories().getPoi().toPath());
-					} catch (IOException ex) {
-						Debug.dumpException(String.format("failed to copy file %s to %s", source, getRegionDirectories().getPoi()), ex);
+					if (source.exists()) {
+						try {
+							Files.copy(source.toPath(), getRegionDirectories().getPoi().toPath());
+						} catch (IOException ex) {
+							Debug.dumpException(String.format("failed to copy file %s to %s", source, getRegionDirectories().getPoi()), ex);
+						}
 					}
 				} else {
 					allCopied = false;
@@ -214,21 +224,22 @@ public class ChunkImporter {
 
 				if (!getRegionDirectories().getEntities().exists() && sourceDirs.getEntities() != null) {
 					File source = new File(sourceDirs.getEntities(), getRegionDirectories().getLocationAsFileName());
-					try {
-						Files.copy(source.toPath(), getRegionDirectories().getEntities().toPath());
-					} catch (IOException ex) {
-						Debug.dumpException(String.format("failed to copy file %s to %s", source, getRegionDirectories().getEntities()), ex);
+					if (source.exists()) {
+						try {
+							Files.copy(source.toPath(), getRegionDirectories().getEntities().toPath());
+						} catch (IOException ex) {
+							Debug.dumpException(String.format("failed to copy file %s to %s", source, getRegionDirectories().getEntities()), ex);
+						}
 					}
 				} else {
 					allCopied = false;
 				}
-			}
 
-			if (!allCopied) {
-				progressChannel.incrementProgress(getRegionDirectories().getLocationAsFileName());
-				return;
+				if (allCopied) {
+					progressChannel.incrementProgress(getRegionDirectories().getLocationAsFileName());
+					return;
+				}
 			}
-
 
 			// ---------------------------------------------------------------------------------------------------------
